+++ CHANGELOG.md
@@ -4,12 +4,10 @@ All notable changes to this project will be documented in this file.
 Version changes are pinned to SDK releases.
 
 ## [Unreleased]
+
 - general: Specify buffer-layout version in package.json. ([#119](https://github.com/zetamarkets/sdk/pull/119))
 - program: Add new NoError cancel instructions. ([#114](https://github.com/zetamarkets/sdk/pull/114))
 - client: Add new instructions ReplaceByClientOrderId, CancelAllOrdersNoError and CancelMultipleOrdersNoError. ([#114](https://github.com/zetamarkets/sdk/pull/114))
-
-
-- events: Add isTaker boolean to PlaceOrderEvent. ([#116](https://github.com/zetamarkets/sdk/pull/116))
 
 ## [0.14.1]